+++ release/src/router/shared/bcmtimer.h
@@ -21,8 +21,8 @@
 #include <time.h>
 
 /* timer ID */
-typedef unsigned int bcm_timer_module_id;
-typedef unsigned int bcm_timer_id;
+typedef unsigned long bcm_timer_module_id;
+typedef unsigned long bcm_timer_id;
 
 /* timer callback */
 typedef void (*bcm_timer_cb)(bcm_timer_id id, int data);
